Design change in App and routes file
It is now permitted to have no routes file. This happens if no filename
was passed in and nothing exists at the default location. In this case,
the app will be loaded with the default Minima::Controller with a single
route setup for '/', to emit a "hello, world".

This is possible thanks to a new method on Minima::Router, which exposes
the connect method of the internal router.

Tests were updated accordingly.